Prefabs, Radio and Printer mechanics
Turned Player, Virtual Cameras and a placeholder for paper into Prefabs,
Created a Station script for all stations scripts to inherit, mainly used for override methods like Activate(), Deactivate and IsZoomer(). Radio currently uses buttons to change the Amp and Frequency values which if the values are close enough to the Target values will result in Successful print when activating the Printer.